# coinstats, assumeutxo: fix hash_serialized2 calculation