@dergoegge 2023-10-19
I've written a fuzz test that catches the original bug mention in https://github.com/bitcoin/bitcoin/issues/28675#issuecomment-1770389468: https://github.com/dergoegge/bitcoin/blob/2023-10-fuzz-coinstats-hash/src/test/fuzz/coinstats_hash.cpp.

I think it also caught a second issue regarding the values of UTXOs, certain negative values seem to result in the same hash as their positive counter part. I have not investigated further but I assume it is realted to `VarIntMode::NONNEGATIVE_SIGNED`.

```bash
$ echo "CMoC9Q==" | base64 --decode > coinstats_hash.crash
$ FUZZ_PRINTF=1 FUZZ=coinstats_hash ./src/test/fuzz/fuzz coinstats_hash.crash
INFO: Running with entropic power schedule (0xFF, 100).
INFO: Seed: 3663010564
INFO: Loaded 1 modules   (379542 inline 8-bit counters): 379542 [0x560080cb5d88, 0x560080d1281e),
INFO: Loaded 1 PC tables (379542 PCs): 379542 [0x560080d12820,0x5600812dd180),
./src/test/fuzz/fuzz: Running 1 inputs 1 time(s) each.
Running: coinstats_hash_crashes/crash-d59c557ae7f9e837716b272a05d1397a94f589d7
2023-10-19T11:34:50Z Opened LevelDB successfully
2023-10-19T11:34:50Z Using obfuscation key for : 0000000000000000
2023-10-19T11:34:50Z mutating a coin
2023-10-19T11:34:50Z coinbase=1 height=1386179375 value=-5934962493229608344 scriptPubKey=[97951910994b0a6104092fbc9b385639]
2023-10-19T11:34:50Z coinbase=1 height=1386179375 value=6856422276773357632 scriptPubKey=[97951910994b0a6104092fbc9b385639]
fuzz: test/fuzz/coinstats_hash.cpp:135: void coinstats_muhash_fuzz_target(FuzzBufferType): Assertion `mutated_result->hashSerialized != result->hashSerialized' failed.
```

Using the varints in the serialization for the hashes seems like unnecessary complexity in the first place. Since we are changing the hash format already, maybe we just get rid of the varints entirely?

(I've not polished the fuzz test, sorry if you run into bugs in the test itself)