@fjahr 2023-10-19
@dergoegge Awesome to have a fuzz test already :)

I think the issue with negative amounts is good to know but not as critical. Since negative output values are not consensus valid it seems ok for the serialization to assume that they are not negative. That's at least how I understand the issue right now. So I think in practice this should at least not lead to a different hash.

I am unsure about removing VARINT from this code entirely. I agree that for the hashing it is not needed but we also use it when we dump the utxo set and there we probably don't want to get rid of it since it's saving space. So we still have to reason about it anyway and I think VARINT isn't really a problem in general. It's widely used in our code and I think also pretty well understood. But I won't fight it if people agree that this is a valuable change.

Let's discuss this in the IRC meeting today!