@fanquake 2023-10-19
I think this should also change `hash_serialized_2` -> `hash_serialized_3`, and add a release note. Given that this was broken, and likely had very few users, I don't see a problem with a breaking change, and making it clear to anyone that may have been using it, that things were broken (better than silently starting to return different results).

I don't see any value in keeping around a broken `hash_serialized_2` either, as that will just lead to future confusion, in relation to AssumeUTXO.