@fjahr 2023-10-19
> I think this should also change `hash_serialized_2` -> `hash_serialized_3`

Makes sense, added