@dergoegge 2023-10-19
> Since negative output values are not consensus valid it seems ok for the serialization to assume that they are not negative.

Couldn't an attacker distribute utxo sets with the negative values instead of the correct positive ones without it being detected at startup? This would then later lead to a consensus split when these outputs are spent.