@jamesob 2023-10-19
> Couldn't an attacker distribute utxo sets with the negative values instead of the correct positive ones without it being detected at startup? This would then later lead to a consensus split when these outputs are spent.

On that note, may not be a bad idea to add a negative-amount sanity check to `ActivateSnapshot()` as a belt-and-suspenders, since we're already checking the height for overflow anyway: https://github.com/jamesob/bitcoin/blob/447f50e4aed9a8b1d80e1891cda85801aeb80b4e/src/validation.cpp#L4926-L4928.