@theStack 2023-10-19
Here are some benchmark results of different branches that has been proposed so far.
I've ran `$ time ./src/bitcoin-cli gettxoutsetinfo` on master, around block height 812940 three times each:

| master (buggy) | this PR (with VARINTs, simple fix, [still buggy](https://github.com/bitcoin/bitcoin/pull/28685#pullrequestreview-1688224112)) | [dergoegge's branch](https://github.com/dergoegge/bitcoin/tree/2023-10-fuzz-coinstats-hash) (without VARINTs, bugfixed) | [my branch](https://github.com/theStack/bitcoin/tree/fix_hash_serialized2_with_muhash_serialization) (using `TxOutSer` serialization, as used in MuHash) |
|----------------|-------------------------------------|------------------------------------------------|---------------------------------------------------------------|
| 1m8,664s       | 1m9,024s                            | 1m9,861s                                       | 1m25,875s                                                     |
| 1m6,803s       | 1m8,815s                            | 1m12,192s                                      | 1m28,743s                                                     |
| 1m8,808s       | 1m9,466s                            | 1m12,227s                                      | 1m28,711s                                                     |

The `TxOutSer` is the simplest one (`ApplyHash` follows the exact same logic as for MuHash, only that the data is fed into `HashWriter` rather than `MuHash3072` instance), but looking at the numbers, it seems that going with @dergoegge's solution makes the most sense?