@fjahr 2023-10-19
I have similar results as @theStack . Involvement of `VARINT` doesn't seem to have much impact on performance, but MuHash style serialization does though that would allow for nice simplifications in coinstats (see that commit).

| master | `VARINT` removed (https://github.com/fjahr/bitcoin/commit/0431c115904b8df344efc864e1b2cc151cafa0ee) | MuHash style (https://github.com/fjahr/bitcoin/commit/6b3628f2c8402695af970a6f2e54da53a30d3c65) |
| 1m23.825s | 1m15.944s | 1m37.636s |
| 1m15.484s | 1m16.271s | 1m39.447s |
| 1m16.042s | 1m16.467s | 1m36.454s |