@fjahr 2023-10-19
> @theStack @fjahr Couldn't the difference in performance be due to the fact that the muhash style version constructs a temporary `DataStream` object first, while the other feeds it to a hasher directly? It's possible to use muhash-style serialization without the intermediary object I think.

Seems that you are right, I have tested this with the [code here](https://github.com/bitcoin/bitcoin/commit/55884a01e2ed08c37879d359c092a555e5038dbf) and I get similar performance as with master (m17.516s, 1m20.996s, 1m22.714s).